refactor to use non-blocking utility for id validations
